==> notebooks/python/01-Lectura_datos_ICC.ipynb <==
## Lectura de Datos 

##### Este notebook tiene como objetivo leer cada una de las hojas del excel de control de calidad de las estaciones meteorologicas de las que esta a cargo ICC

pip install openpyxl


## Libreías generales
import numpy as np
import pandas as pd
import datetime
from dateutil.relativedelta import relativedelta
import openpyxl
import os

#comando para visualizar todas las columnas del df
pd.pandas.set_option('display.max_columns', None)

os.chdir('D:/OneDrive - CGIAR/Proyectos/AgriLAC-ICC/ICC/Scrip_bases/Carpeta_compartir')

###Carga del excel donde se encuentra el control de calidad de cada una de las estaciones

data_control= pd.read_excel("control_calidad.xlsx", sheet_name=None)
data_control.keys()
print(data_control)

### Identificar cada hoja de excel con el nombre de la estacion

datos=pd.DataFrame()
diccionario=pd.read_excel("control_calidad.xlsx",sheet_name=None)


for nombre,columna in diccionario.items():
    columna['estacion']=nombre
    datos=datos.append(columna, ignore_index=True)
    

datos=datos[['estacion','Fecha', 'temperatura', 'radiacion', 'humedad relativa',
       'precipitacion', 'velocidad viento', 'mojadura', 'direccion viento','presion atmosferica']]

datos = datos.rename(columns={'Fecha':'fecha','velocidad viento':'velocidad_viento','humedad relativa':'humedad_relativa',
                                   'direccion viento':'direccion_viento', 'presion atmosferica': 'presion_atm'})

"""
Se crea un diccionario con las abreviaturas de los nombres de cada estación

"""

dict_esta= {'CEN':"CEN-CEN",'TEH':"LUT-TEH", 'BAL':"PAN-BAL", 'PUY':"TBU-PUY", 'SAV':"MAG-SAV", 'AMA':"SAA-AMA",
             'TRI':"SDT-TRI", 'IRL':"TBU-IRL", 'BON':"LUT-BON", 'BOU':"MAG-BOU", 'NRJ':"PAG-NRJ", 'PEO':"MAT-PEO",
             'CBR':"SDT-CBR", 'TLA': "TUL-TLA", 'LOR':"PAG-LOR" , 'SRF':"PAN-SRF", 'XOL':"MAG-XOL", 'SNC':"MAG-SNC",
             'TMG':"MAG-TMG", 'PLT':"ICC-PLT",  'LGR':"SAA-LGR", 'CDL':"ICC-CDL",  'CHI':"ICC-CHI", 'PVD':"MAG-PVD", 
             'ALA':"ICC-ALA", 'COC':"PAG-COC", 'CON':"ICC-CON", 'LMQ':"SAA-LMQ", 'STA':"ICC-STA", 'YEPO':"ICC-YEPO",
             'CIZ':"CASSA-CIZ",'CPL':"CASSA-CPL", 'CPT':"ASSA-CPT", 'ECA':"CASSA-ECA", 'SCL':"CASSA-SCL", 'AGS':"CASSA-AGS",
             'PBR':"ICC-PBR", 'MTA':"ICC-MTA", 'TUL': "TUL-TLA", 'NAR': "PAG-NRJ", 'LCD':"ICC-CDL", 'CASSA-CPT':"ASSA-CPT", 
             'SOTZ':"'ICC-SOTZ", 'JOY':"ICC-JOY", 'PUY':"TBU-PUY", 'CTB':"SDT-CBR" , 'SRAF':"PAN-SRF"
}


"""
Se realiza el reemplazo de las abrevituras a partir del diccionario creado

"""
datos.replace({"estacion": dict_esta},inplace=True)

###### Se crea la columna de tiempo
datos['Hora']=''
contador=0
for i in range(len(datos)):
    if contador==1440:
        contador=0
    Hora=int(contador/60)
    Minutos=contador%60
    Tiempo=str(Hora)+':'+ str(Minutos)
    datos.at[i,'Hora']=Tiempo
    contador+=15


###### se transforma la columna fecha en caracter
datos['fecha']= datos.fecha.astype(str)
#### Se coloca en formato tiempo la columna Hora
datos['Hora']=pd.to_datetime(datos['Hora'], format='%H:%M')
#### Separo la hora de la fecha
datos['new_time'] = [d.time() for d in datos['Hora']]
##Elimino la columna ora
datos=datos.drop(['Hora'],axis=1)
#### la columna de tiempo la transformamos en caracter
datos['new_time']=datos.new_time.astype(str)
### En la columna del tiempo se toman los primero 5 cacateres que hacen referencia a la hora y minutos
datos["new_time"] = datos["new_time"].apply(lambda x: x[0:5])
### Se concatena la columna de fecha con el tiempo y se elimina la columna de tiempo, obteniendo así la data final
datos['fecha'] = datos.fecha.str.cat(datos.new_time, sep=' ')
datos=datos.drop(['new_time'],axis=1)


def date_convert(date_to_convert):
     
     return datetime.datetime.strptime(date_to_convert, "%Y-%m-%d %H:%M").strftime("%Y-%d-%m %H:%M")


datos['fecha'] = datos['fecha'].apply(date_convert)


datos.head()

#### dirección donde se ubicaran los archivos de las estaciones
os.chdir('D:/OneDrive - CGIAR/Proyectos/AgriLAC-ICC/ICC/Scrip_bases/Carpeta_compartir/estaciones_control')


### Guardar cada estacion es un archivo de excel diferente 

for i in datos.estacion.unique():
    
    df2=datos[datos['estacion']==i]
    df2.to_csv(f"{i}.csv",index=False, header=False)
