==> notebooks/python/07-Boletin _climatico_semanal_ICC.ipynb <==
## Boletín climático semanal ICC

##### Este notebook tiene como objetivo construir la tabla de reporte que realiza ICC

## Libreías generales
import numpy as np
import pandas as pd
import datetime
from dateutil.relativedelta import relativedelta

import matplotlib.pyplot as plt
from matplotlib import cm
import matplotlib

# Configuración warnings
# ==============================================================================
import warnings
warnings.filterwarnings('ignore')

#comando para visualizar todas las columnas del df
pd.pandas.set_option('display.max_columns', None)



### Lectura de archivo diario 

dfs_diario=pd.read_parquet("diario_complete.parquet")

### Lectura de archivo semanal

dfs_semanal=pd.read_parquet("semanal_complete.parquet")

### Lectura de archivo anual

dfs_anual=pd.read_parquet("anual_complete.parquet")
dfs_anual.head()


dfs_semanal_agg= dfs_diario.groupby(['estrato','semana','año']).agg(

    hume_rela_min_semanal= ('humedad_relativa_min_diaria','mean'),
    hume_rela_max_semanal= ('humedad_relativa_max_diaria','mean'),
    hume_rela_media_semanal=('humedad_relativa_media_diaria','mean'),
    ETP_media= ('FAO-Penman-Montieth','mean')
    

).reset_index().round(2).astype({'semana':int, 'año':int})

"""
Se cruza la base con los periodos seleccionados junto con las nuevas variables 
"""
dfs_semanal=dfs_semanal.merge(dfs_semanal_agg, how='right', on=['estrato','semana','año'])


Semana= int(input("Digite la semana que desea visualizar"))

week_report = dfs_semanal[dfs_semanal['semana'] == Semana]
recent_year = week_report['año'].max()
sele_period = list(range(recent_year, recent_year - 4, -1))
week_report = week_report[week_report['año'].isin(sele_period)]

Boletin=['estrato','semana', 'año','temp_min_semanal', 'temp_promedio_semanal',
         'temp_max_semanal','ampl_term_prom_semanal','hume_rela_min_semanal',
         'hume_rela_media_semanal','hume_rela_max_semanal','rad_diar_prom_semanal',
         'n_brillo_solar_acum','preci_semana','prec_acumulad_anual','Veloc_viento_media_semanal','Veloc_viento_prom_max_semanal','ETP_media']

week_report=week_report[Boletin]

week_report.to_excel('tabla_boletin_climatico_semanal.xlsx')
