==> notebooks/python/08_Boletin_gremio_bananero.ipynb <==
## Boletin semanal asociación de productores independientes de banano

##### Este notebook tiene como objetivo construir la tabla y graficos de reporte que realiza ICC

## Libreías generales
import numpy as np
import pandas as pd
import datetime
from dateutil.relativedelta import relativedelta

# Gráficos
# ==============================================================================
import matplotlib.pyplot as plt
import matplotlib.font_manager
from matplotlib import style
style.use('ggplot') or plt.style.use('ggplot')
import seaborn as sns

import matplotlib.pyplot as plt
from matplotlib import cm
import matplotlib

# Configuración warnings
# ==============================================================================
import warnings
warnings.filterwarnings('ignore')

#comando para visualizar todas las columnas del df
pd.pandas.set_option('display.max_columns', None)

### Lectura de archivo diario 

dfs_diario=pd.read_parquet("diario_complete.parquet")
dfs_diario.head()

##3Lectura archivo de la información por estación

info_estaciones = pd.read_excel('prueba_info.xlsx')

info_estaciones.head()

###Limpieza de las columna de codigo
info_estaciones.Codigo =info_estaciones.Codigo.replace(r'\W+', '', regex=True)
info_estaciones.Codigo = info_estaciones.Codigo.replace(r'\s+', ' ', regex=True)
info_estaciones.Codigo = info_estaciones.Codigo.replace("", np.nan, regex=True)
info_estaciones.Codigo= info_estaciones.Codigo.replace(" ", np.nan, regex=True)

### Columnas que se necesitan
info_estrato =info_estaciones[['Estacion','Codigo','t','Altitud\n(msnm)','LUGAR']]
info_estrato.columns = ['nombre_estacion','estacion','estrato','Altitud\n(msnm)','Ubicacion']

df_semanal_estacion = dfs_diario.groupby(['estacion','semana','año']).agg(
    temp_min_semanal=('temperatura_min_diaria', 'mean'),
    temp_max_semanal=('temperatura_max_diaria', 'mean'),
    temp_promedio_semanal=('temperatura_promedio_diaria', 'mean'),
    rad_diar_prom_semanal=('radiacion_diaria_promedio', 'mean'),
    ampl_term_prom_semanal = ('amplitud_termica', 'mean'),
    hume_rela_med_semanal= ('humedad_relativa_media_diaria','mean'),
    Veloc_viento_prom_max_semanal= ('velocidad_viento_max_diaria', 'mean'),
    Veloc_viento_media_semanal= ('velocidad_viento_media_diario', 'mean' ),
    max_lluvia=('lluvia_diaria', 'max'),
    acumu_lluvia=('lluvia_diaria', 'sum'),
    suma_FAO= ('FAO-Penman-Montieth','sum')

).reset_index().round(2).astype({'semana':int, 'año':int})

df_semanal_estacion

df_semanal_estacion['Balance_precipitacion-evapotranspiracion']=df_semanal_estacion.acumu_lluvia-df_semanal_estacion.suma_FAO

df_semanal_estacion=df_semanal_estacion.merge(info_estrato, how='right', on=['estacion'])
df_semanal_estacion.head()

Boletin_banano=df_semanal_estacion[['nombre_estacion','Altitud\n(msnm)','Ubicacion','semana', 'año','temp_min_semanal','temp_promedio_semanal','temp_max_semanal','hume_rela_med_semanal',
                                    'acumu_lluvia','suma_FAO','Balance_precipitacion-evapotranspiracion']]
Boletin_banano.columns = ['Estación','Altitud(msnm)','Ubicacion','semana', 'año','temp_min_semanal','temp_promedio_semanal','temp_max_semanal','hume_rela_med_semanal',
                                    'Precipitacion acumulada','Evapotranspiración_acumulada','Balance_precipitacion-evapotranspiracion']

"""
se elimnan las estaciones 'PLT':"ICCPLT", 'STA':"ICCSTA", 'PBR':"ICCPBR" y 'YEPO':"ICCYEPO", 
dado que los reportes que realizan desde ICC solo incluyen las estaciones de zonas cañeras, 
por lo tanto se reportan 28 estaciones 

"""

to_remove = Boletin_banano.loc[Boletin_banano.Estación.isin(['Santa Teresa ', 'FCA Yepocapa', 'Central Izalco ', 'Copal ', 'Chaparrastique', 'El Carmen', 'San Clemente', 'Agua Santa', 'Pueblo Real ','El Platanar'])].index

Boletin_banano.drop(to_remove, inplace=True)

Semana= int(input("Digite la semana que desea visualizar"))
periodo=int(input("Digite el año que desea visualizar"))

week_banano=Boletin_banano[(Boletin_banano.semana==Semana)&(Boletin_banano.año==periodo)]
week_banano.head(28)

### Guardar tabla en excel 
week_banano.to_excel('boletin_banano.xlsx')
